% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSeqTaxonomy.R
\name{assignSeqTaxonomy}
\alias{assignSeqTaxonomy}
\title{Assign Taxonomy
Uses DADA2's assignTaxonomy function to assign taxa to ASV sequences.}
\usage{
assignSeqTaxonomy(seqtab, object)
}
\arguments{
\item{seqtab}{Sequence frequency table}

\item{object}{S4 object of type fastq2otu (i.e. fastPaired, fastSingle)}
}
\value{
OTU table
}
\description{
Assign Taxonomy
Uses DADA2's assignTaxonomy function to assign taxa to ASV sequences.
}
