% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_assign_tax}
\alias{check_assign_tax}
\title{Check for valid inputs for fastAssignTaxa}
\usage{
check_assign_tax(object)
}
\arguments{
\item{object}{fastAssignTaxa-type S4 object}
}
\value{
TRUE (boolean) or error message detailing source(s) of error.
}
\description{
This function checks to see if fastAssignTaxa parameters are present and valid. Missing or invalid inputs will cause an
error message to be displayed.
}
