% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_fastPaired}
\alias{check_fastPaired}
\title{Check for valid inputs for fastPaired}
\usage{
check_fastPaired(object)
}
\arguments{
\item{object}{fastPaired-type S4 object}
}
\value{
TRUE (boolean) or error message detailing source(s) of error.
}
\description{
This function checks to see if the isPaired parameter reflects the object the user is attempting to create.
Generates an additional level of validity that insures uses is using package correctly.
}
