% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_fastq2otu}
\alias{check_fastq2otu}
\title{Check for valid inputs for fastq2otu}
\usage{
check_fastq2otu(object)
}
\arguments{
\item{object}{fastq2otu-type S4 object}
}
\value{
TRUE (boolean) or error message detailing source(s) of error.
}
\description{
This function checks to see if user inputs are present and valid. Missing or invalid inputs will cause an
error message to be displayed.
}
