% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_primer_trim}
\alias{check_primer_trim}
\title{Check for valid inputs for fastPrimerTrim}
\usage{
check_primer_trim(object)
}
\arguments{
\item{object}{fastPrimerTrim-type S4 object}
}
\value{
TRUE (boolean) or error message detailing source(s) of error.
}
\description{
This function checks to see if "pathToBBduk", "listOfAdapters", "pathToRawFastq" and "pathToNoPrimers" parameters are present and valid. Missing or invalid inputs will cause an
error message to be displayed.
}
