% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_seq_dump}
\alias{check_seq_dump}
\title{Check for valid inputs for fastSeqDump}
\usage{
check_seq_dump(object)
}
\arguments{
\item{object}{fastSeqDump-type S4 object}
}
\value{
TRUE (boolean) or error message detailing source(s) of error.
}
\description{
This function checks to see if fastSeqDump parameters are present and valid. Missing or invalid inputs will cause an
error message to be displayed.
}
