% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadaSeqs.R
\name{dadaSeqs}
\alias{dadaSeqs}
\title{Denoise sequence data}
\usage{
dadaSeqs(derep, err, object)
}
\arguments{
\item{derep}{Output(s) of derepSeqs function}

\item{err}{Output(s) of learnErrors function}
}
\value{
dada object
}
\description{
Denoise sequence data
}
