% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastAssignTaxa-class}
\alias{fastAssignTaxa-class}
\title{FASTASSIGNTAXA Custom Class}
\description{
Use to set parameters for DADA2 assignTaxonomy function.
}
\details{
Sub-class of fastq2otu S4 datatype.
}
