% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastFilter-class}
\alias{fastFilter-class}
\title{FASTFILTER Custom Class}
\description{
Use to set parameters for DADA2 filterAndTrim function. Inputs for maxEE, truncQ, truncLen, and
trimming functions (trimLeft, trimRight) are validated by check_filt_params function.
}
\details{
Sub-class of fastSingle and fastPaired S4 datatypes.
}
