% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastPlotQuality-class}
\alias{fastPlotQuality-class}
\title{FASTPLOTQUALITY Custom Class}
\description{
Use to set parameters for DADA2's plotQualityProfile function.
}
\details{
Sub-class of fastq2otu S4 datatype.
}
