% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastPrimerTrim-class}
\alias{fastPrimerTrim-class}
\title{FASTPRIMERTRIM Custom Class}
\description{
Use to set parameters for removing adapter sequences from amplicons using BBTools' bbduk.sh
script.
}
\details{
Sub-class of fastSingle and fastPaired S4 datatypes.
}
