% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastq2otu-class}
\alias{fastq2otu-class}
\title{FASTQ2OTU Custom Super-Class}
\description{
Use to convert parameters provided by user to accessible attributes associated with
an S4-type data class.
}
