% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastReport-class}
\alias{fastReport-class}
\title{FASTREPORT Custom Class}
\description{
Use to set parameters for generating fastqc reports using FASTQCR.
Requires FASTQC to be installed onto system.
}
\details{
Sub-class of fastq2otu S4 datatype.
}
