% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastSeqDump-class}
\alias{fastSeqDump-class}
\title{FASTSEQDUMP Custom Class}
\description{
Use to set parameters for retrieving FASTQ files from NCBI's SRA database
}
\details{
Sub-class of fastq2otu S4 datatype.
}
