% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{fastSingle-class}
\alias{fastSingle-class}
\title{Inherits methods and attributes from fastq2otu class
Lists paramaters that are essential for analysing single-end data}
\description{
Inherits methods and attributes from fastq2otu class
Lists paramaters that are essential for analysing single-end data
}
