% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtTrim.R
\name{filtTrim}
\alias{filtTrim}
\title{Filter and trim single or paired-end data}
\usage{
filtTrim(object, sample.names)
}
\arguments{
\item{object}{S4 object of type class fastq2otu (i.e. fastq2otu_single, fastq2otu_paired)}
}
\description{
Filter and trim single or paired-end data
}
