% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanQualityProfile.R
\name{getMeanQualityProfile}
\alias{getMeanQualityProfile}
\title{Get quality profile of a fastq file(s).}
\usage{
getMeanQualityProfile(fl, n = 5e+05, project, layout = "Single")
}
\arguments{
\item{fl}{(Required). \code{character}.
File path(s) to fastq or fastq.gz file(s).}

\item{n}{(Required). Default 500,000.
The number of records to sample from the fastq file.}

\item{project}{(Required). Provide a character label that describes the data being handled.}

\item{layout}{(Required). Default is Single.
Provide "Single" if handling single-end data; "Forward" if handling foward-reads,
and "Reverse" if handling reverse-reads.}
}
\value{
A table with the following column names: c("Position", "Mean", "Orientation", "Study")
}
\description{
This function generates a table that reports the mean quality scores
observed at each sequence position for the input fastq file(s).
}
\details{
Code modified from DADA2's plotQualityProfile function
}
