% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnSeqErrors.R
\name{learnSeqErrors}
\alias{learnSeqErrors}
\title{Learn Sequence Errors
Enables exection of DADA2's learnErrors functions.
Multithread and verbose parameters are true by default}
\usage{
learnSeqErrors(file)
}
\arguments{
\item{file}{path to FASTQ file}
}
\value{
error object
}
\description{
Learn Sequence Errors
Enables exection of DADA2's learnErrors functions.
Multithread and verbose parameters are true by default
}
