% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSeqsTable.R
\name{makeSeqsTable}
\alias{makeSeqsTable}
\title{Generate Sequence Frequency Table
Uses DADA2 makeSequenceTable function to generate a sequence (ASV) frequency table
that provides the number of reads per ASV.}
\usage{
makeSeqsTable(dadaObj, sample.name)
}
\arguments{
\item{dadaObj}{Object returned from mergePairs or dada functions}
}
\value{
matrix with 1-2 rows. ASVs are listed as column names and each row contains frequency information belonging to a single sample.
}
\description{
Generate Sequence Frequency Table
Uses DADA2 makeSequenceTable function to generate a sequence (ASV) frequency table
that provides the number of reads per ASV.
}
