% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSamples.R
\name{mergeSamples}
\alias{mergeSamples}
\title{Merge OTU tables from different samples}
\usage{
mergeSamples(otutabs, seqtabs, label, taxLevels)
}
\arguments{
\item{otutabs}{List of path(s) to OTU tables}

\item{seqtabs}{List of path(s) to frequency tables}

\item{final.print}{Name of file to save.}
}
\value{
Merged table that can be used to complete cross sample comparisons
}
\description{
Merge OTU tables from different samples
}
