% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuality.R
\name{plot_quality}
\alias{plot_quality}
\title{Create a function the plots quality graphs on individual pages within a .pdf file}
\usage{
plot_quality(data, n)
}
\arguments{
\item{data}{Path(s) to input files}

\item{n}{Sampling number}
}
\description{
Create a function the plots quality graphs on individual pages within a .pdf file
}
