% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuality.R
\name{plotQuality}
\alias{plotQuality}
\title{plotQuality}
\usage{
plotQuality(fp, label, object)
}
\arguments{
\item{fp}{Path to directory containing FASTQ files}

\item{label}{Unique file or project id used to generate file/plot title}

\item{object}{fastPlotQuality object}
}
\description{
plotQuality
}
