% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeqOverlap.R
\name{plotSeqOverlap}
\alias{plotSeqOverlap}
\title{Plot the quality profile of paired-end data
The figure created can be used to visually determine the best merging and trimming parameters.}
\usage{
plotSeqOverlap(
  tabF,
  tabR,
  primerF,
  primerR,
  projectLabel = "Paired-end Study",
  matchScore = 1,
  mismatchScore = -3,
  gapOpeningScore = 5,
  gapExtensionScore = 2,
  minScore = 4
)
}
\arguments{
\item{tabF}{(Required). Forward table obtained from the getMeanQualityProfile() function.}

\item{tabR}{(Required). Reverse table obtained from the getMeanQualityProfile() function.}

\item{primerF}{(Required). Forward primer sequence provided in the 5' to 3' orientation (must NOT feature any degenerate bases or invalid
characters).}

\item{primerR}{(Required). Reverse primer sequence provided in the 5' to 3' orientation (must NOT feature any degenerate bases or invalid
characters).}

\item{projectLabel}{(Required). Default is "Paired-end Study".
Character label (i.e. project accession number) that can be used to describe data being analyzed.}

\item{matchScore}{(Default is 1). Match alignment score.}

\item{mismatchScore}{(Default is -3). Mismatch alignment score.}

\item{gapOpeningScore}{(Default is 5). Gap opening score.}

\item{gapExtensionScore}{(Default is 2). Gap extension score.}

\item{minScore}{(Default is 10). Minimum alignment score to accept.}
}
\value{
ggplot figure
}
\description{
Plot the quality profile of paired-end data
The figure created can be used to visually determine the best merging and trimming parameters.
}
