% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_readConfig.R, R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{Read in a config file (preferred method)}
\usage{
readConfig(configFile, isPaired = FALSE, type = c("auto"))

readConfig(configFile, isPaired = FALSE, type = c("auto"))
}
\arguments{
\item{configFile}{Path to config file (YML-formatted)}

\item{isPaired}{Default is FALSE. If TRUE, parameters for paired-end data will be used.}

\item{type}{Default is 'auto'. String or vector containing the following value(s): 'auto', 'assignTax', 'report', 'seqdump', 'primertrim', 'filter', 'qualityplot'. Used to specify what type of object is being created.
If a vector is provided it must contain one of the following combination of strings c('auto', 'assignTax') and  c('auto', 'filter')}
}
\value{
An S4 object

An S4 object
}
\description{
This function parses config file to extract data needed to create a FASTQ2OTU class object. It assumes that
the file is in YML format and contains default key terms.
Execute getSlots("fastq2otu") to get all required arguments or refer to documenation for more details and/or
examples.

This function parses config file to extract data needed to create a FASTQ2OTU class object. It assumes that
the file is in YML format and contains default key terms.
Execute getSlots("fastq2otu") to get all required arguments or refer to documenation for more details and/or
examples.
}
\details{
Use when config file is supplied and all required values are provided.

Use when config file is supplied and all required values are provided.
}
