% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeChimeras.R
\name{removeChimeras}
\alias{removeChimeras}
\title{Remove Chimeric Sequences
Enables the use of DADA2's isBimeraDenovoTable and/or removeBimeraDenovo functions.}
\usage{
removeChimeras(seqtab, object)
}
\arguments{
\item{seqtab}{Frequency table(s) returned from makeSeqsTable function}
}
\description{
Remove Chimeric Sequences
Enables the use of DADA2's isBimeraDenovoTable and/or removeBimeraDenovo functions.
}
