% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastqc.R
\name{runFastqc}
\alias{runFastqc}
\title{Generate FASTQC Report}
\usage{
runFastqc(object, path)
}
\arguments{
\item{object}{fastq2otu-type class object}

\item{path}{Path to directory containing input fastq files}
}
\description{
Generate FASTQC Report
}
