% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_runPipeline.R, R/modified_runPipeline.R,
%   R/runPipeline.R
\name{runPipeline}
\alias{runPipeline}
\title{Execute project workflow from FASTQ input to OTU output using DADA2 Workflow}
\usage{
runPipeline(
  configFile,
  isPaired = FALSE,
  getQuality = TRUE,
  getMergedSamples = TRUE,
  getDownloadedSeqs = FALSE,
  getTrimmedAdapters = FALSE,
  getGeneratedReport = FALSE
)

runPipeline(
  configFile,
  isPaired = FALSE,
  getQuality = TRUE,
  getMergedSamples = TRUE,
  getDownloadedSeqs = FALSE,
  getTrimmedAdapters = FALSE,
  getGeneratedReport = FALSE
)

runPipeline(
  configFile,
  isPaired = FALSE,
  getQuality = TRUE,
  getMergedSamples = TRUE,
  getDownloadedSeqs = FALSE,
  getTrimmedAdapters = FALSE,
  getGeneratedReport = FALSE
)
}
\arguments{
\item{configFile}{Path to config file (YML-formatted).}

\item{isPaired}{Default if FALSE. If TRUE workflow for paired end data is executed.}

\item{plotQuality}{Default is TRUE. Sequence quality distribution is plotted using DADA2's plotQualityProfile method.}

\item{getMergeSamples}{Default is TRUE. If FALSE, sample OTU tables will not be merged across.}

\item{downloadSeqs}{Default is FALSE. If TRUE, users will be able to retrieve FASTQ files from SRA using the fastq2otu's getSeqs method.}

\item{trimAdapters}{Default is FALSE. If TRUE, users will be able to remove adapters sequences from data using BBTools bbduk.sh script.}

\item{generateReport}{Default is FALSE. If TRUE, allows a FASTQC report to be generated from input data.}
}
\description{
The analysis takes place in multiple steps beginning at the creation of a central output directory. Once the directory
is created, a log-file is initilized that will contain all messages produced by DADA2. Accompanying the log file is a summary
table documenting changes in read frequency following filtering and trimming procedures. This summary table can be used
as a reference when determining the best way to modify different parameters input into the different functions.

The analysis takes place in multiple steps beginning at the creation of a central output directory. Once the directory
is created, a log-file is initilized that will contain all messages produced by DADA2. Accompanying the log file is a summary
table documenting changes in read frequency following filtering and trimming procedures. This summary table can be used
as a reference when determining the best way to modify different parameters input into the different functions.

The analysis takes place in multiple steps beginning at the creation of a central output directory. Once the directory
is created, a log-file is initilized that will contain all messages produced by DADA2. Accompanying the log file is a summary
table documenting changes in read frequency following filtering and trimming procedures. This summary table can be used
as a reference when determining the best way to modify different parameters input into the different functions.
}
