% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSeqs.R
\name{saveSeqs}
\alias{saveSeqs}
\title{Save sequence frequency tables in RDS file}
\usage{
saveSeqs(table, sample.name, index, output.dir, add.table = FALSE, label)
}
\arguments{
\item{output.dir}{Output directory to write RDS file}

\item{add.table}{Generates an additional table storing all chimeric sequences flagged by DADA2's}

\item{tables}{Sequence frequency table.}

\item{sample.names}{SRA/fastq file ids (extracted directly from file)}
}
\description{
Save sequence frequency tables in RDS file
}
