% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTaxonomyTables.R
\name{saveTaxonomyTables}
\alias{saveTaxonomyTables}
\title{Save OTU Tables}
\usage{
saveTaxonomyTables(table, sample.name, output.dir, index, label)
}
\arguments{
\item{tables}{List of  OTU tables}

\item{sample.names}{List of file ids}
}
\value{
Path(s) to saved OTU tables
}
\description{
Saves list of tables as tab-delimited CSV files
}
