% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setFastFilter}
\alias{setFastFilter}
\title{Create a fastFilter object}
\usage{
setFastFilter(
  inDir,
  outDir,
  prefix = "myproject",
  maxEE = 2.5,
  truncQ = 0,
  truncLen = 0,
  trimLeft = 0,
  trimRight = 0,
  matchIDs = FALSE,
  minLen = 50,
  isPaired = FALSE,
  verbose = FALSE,
  multithread = FALSE
)
}
\arguments{
\item{inDir}{Required input. Path to FASTQ files.}

\item{outDir}{Required input. Path to write output files.}

\item{prefix}{Default is "myproject". Project-specific label that will be included in the names of all files and directories created.}

\item{maxEE}{Default is 2.5. Sets maxEE parameter for DADA2's filterAndTrim method. Two numeric inputs are expected for paired-end data.}

\item{truncQ}{Default is 0. Sets truncQ parameter for DADA2's filterAndTrim method. Two numeric inputs are expected for paired-end data.}

\item{truncLen}{Default is 0. Sets truncLen parameter for DADA2 filterAndTrim function. Two numeric inputs are expected for paired-end data.}

\item{trimLeft}{Default is 0. Sets trimLeft parameter for DADA2 filterAndTrim function. Two numeric inputs are expected for paired-end data.}

\item{trimRight}{Default is 0. Sets trimRight parameter for DADA2 filterAndTrim function. Two numeric inputs are expected for paired-end data.}

\item{matchIDs}{Default is FALSE. Sets matchIDs parameter for DADA2 filterAndTrim function. Set TRUE when handling paired-end data with complementary file names for forward and reverse reads.}

\item{minLen}{Default is 50. Sets minLen parameter for DADA2 filterAndTrim function. Two numeric inputs are expected for paired-end data.}

\item{isPaired}{Default is FALSE. If TRUE, two values must be provided for truncQ, truncLen, trimLeft and trimRight, and minLen (see example below).

TODO: Add Example}

\item{verbose}{Default is FALSE. Sets verbose parameter for DADA2 methods.}

\item{multithread}{Default is FALSE. Sets multithread parameter for DADA2 methods.}
}
\description{
Create a fastFilter object
}
