% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setFastPaired}
\alias{setFastPaired}
\title{Create a fastPaired object}
\usage{
setFastPaired(
  inDir,
  outDir,
  validate = TRUE,
  pattern = ".*_[1,2].fastq(.gz)?",
  mergeSeqs = TRUE,
  trimOverhang = FALSE,
  minOverlap = 12,
  maxMismatch = 0,
  returnRejects = FALSE,
  justConcatenate = FALSE,
  verbose = FALSE,
  prefix = "myproject",
  isPaired = TRUE,
  derepN = 1e+06,
  getErrPDF = FALSE,
  errN = 1e+08,
  multithread = FALSE,
  dadaBandSize = 16,
  dadaOmegaA = 1e-40,
  getChimeraTable = FALSE,
  minSampleFraction = 0.9,
  ignoreNegatives = 1,
  minFoldParentOverAbundance = 1.5,
  detectionAbundance = 2,
  allowOneOff = FALSE,
  maxShift = 16,
  downloadSeqs = FALSE,
  trimAdapters = FALSE,
  generateReport = FALSE
)
}
\arguments{
\item{inDir}{Required input. Path to FASTQ files.}

\item{outDir}{Required input. Path to write output files.}

\item{validate}{Default is TRUE. Execute all validation methods.}

\item{pattern}{Regex pattern used to select fastq files from directory.}

\item{mergeSeqs}{Required input. Default is TRUE to merge forward and reverse amplicons. FALSE prevents DADA2's mergePairs method from being executed.}

\item{trimOverhang}{Default is FALSE. Sets trimOverhang parameter for DADA2's mergePairs method.}

\item{minOverlap}{Default is 12. Sets minOverlap parameter for DADA2's mergePairs function.}

\item{maxMismatch}{Default is 0. Sets maxMismatch parameter for DADA2's mergePairs method.}

\item{returnRejects}{Default is FALSE. Sets returnRejects paramater for DADA2's mergePairs method.}

\item{justConcatenate}{Default is FALSE. Sets returnRejects paramater for DADA2's mergePairs method.}

\item{verbose}{Default is FALSE. Sets verbose parameter for DADA2 methods.}

\item{prefix}{Default is "myproject". Project-specific label that will be included in the names of all files and directories created.}

\item{isPaired}{Default is TRUE for fastPaired object. If FALSE, a fastSingle object will be created instead.}

\item{derepN}{Default is 1e+06. Sets the sampling parameter for DADA2's derepFastq method.}

\item{getErrPDF}{Default is FALSE. If TRUE a learn errors plot from DADA's plotErrors method is produced.}

\item{errN}{Default is 1e+08. Sets the nbases parameter for DADA2's learnErrors method.}

\item{multithread}{Default is FALSE. Sets multithread parameter for DADA2 methods.}

\item{dadaBandSize}{Default is 16. Sets the BAND_SIZE parameter for DADA2's dada method.}

\item{dadaOmegaA}{Default is 1e-40. Sets the OMEGA_A paramter for DADA2's dada method.}

\item{getChimeraTable}{Default is FALSE. When TRUE, creates table showing all sequences labled as chimeric by DADA2 (uses DADA2's isBimeraDevovoTable function as opposed to removeBimeraDenovo).}

\item{minSampleFraction}{Default is 0.9. Sets minDetection parameter for DADA2's removeBimeraDenovo and isBimera functions.}

\item{ignoreNegatives}{Default is 1. Sets ignoreNegatives parameter for DADA2's removeBimeraDenovo and isBimera functions.}

\item{minFoldParentOverAbundance}{Default is 1.5. Sets minfoldParentOverabundance parameter for DADA2's removeBimeraDenovo and isBimera functions.}

\item{detectionAbundance}{Default is 2. Sets detectionParentAbundace parameter for DADA2's removeBimeraDenovo and isBimera functions.}

\item{allowOneOff}{Default is FALSE. Sets allowOneOff parameter for DADA2's removeBimeraDenovo and isBimera functions.}

\item{maxShift}{Default is 16. Sets maxShift parameter for DADA2's removeBimeraDenovo and isBimera functions.}

\item{generateReport}{Default is FALSE. If TRUE, a fastReport object will be created when runPipeline is executed.}

\item{downloadSeqs.}{Default is FALSE. If TRUE, a fastSeqDump object will be created when runPipeline method is executed.}

\item{trimAdapters.}{Default is FALSE. If TRUE, a fastPrimerTrim object will be created when runPipeline method is executed.}
}
\description{
Create a fastPaired object
}
