% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setFastReport}
\alias{setFastReport}
\title{Create a fastReport object
Users have the option of creating a detailed report of their data using FASTQC}
\usage{
setFastReport(
  inDir,
  outDir,
  fastqcPath = NA_character_,
  installFastqc = FALSE,
  numThreads = 4,
  description = NA_character_
)
}
\arguments{
\item{inDir}{Required input. Path to directory containing FASTQ files.}

\item{outDir}{Required input. Path to write output files.}

\item{adapterList}{Required input. List of adapter sequences to remove from files.

TODO: Add Example}
}
\description{
Create a fastReport object
Users have the option of creating a detailed report of their data using FASTQC
}
