% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimAdapters.R
\name{trimAdapters}
\alias{trimAdapters}
\title{Remove primers from sequences}
\usage{
trimAdapters(object)
}
\arguments{
\item{object}{An S4 object of fastPrimerTrim}
}
\value{
file path(s) to trimmed data
}
\description{
Uses DADA2's removePrimers to remove adapter sequences from FASTQ reads.
To use the function only a single parameter is required. However, to use the bbduk script correctly the
fastq2otu object must have its listOfAdapters, listOfSamples, and pathToRawFastq attributes defined.
The listOfAdapters should define the path to a text file containing all adapter sequences to remove.
The listOfSamples should define the path to a text file containing all SRA sample IDs
By default, the trimmed/output files will be written to the path defined in the inDir attribute of the fastq2otu class object.
The path selected for the output files may or may not exist prior to the execution of this function.
To learn more about the fastq2out class and its attributes, please refer back to the documentation.
Creates output directory if it does not already exist
}
