% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSeqs.R
\name{saveSeqs}
\alias{saveSeqs}
\title{Save sequence frequency tables in RDS file}
\usage{
saveSeqs(tables, sample.names, out.dir)
}
\arguments{
\item{tables}{List of sequence frequency tables}

\item{sample.names}{List of SRA/fastq file ids (extracted directly from files)}

\item{out.dir}{Output directory to write RDS files}
}
\description{
Save sequence frequency tables in RDS file
}
