% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setfastPlotQuality}
\alias{setfastPlotQuality}
\title{Create a fastPlotQuality object}
\usage{
setfastPlotQuality(aggregate = FALSE, N = 5e+05, validate = FALSE)
}
\arguments{
\item{aggregate}{Default if FALSE. Sets aggregate parameter for DADA's plotQualityProfile method.}

\item{N}{Default is 5e+05. Sets N parameter for DADA's plotQualityProfile method.}

\item{validate}{Default is FALSE, when TRUE all validation methods are executed.
TODO: Add Example}
}
\description{
Create a fastPlotQuality object
}
