% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setFastPrimerTrim}
\alias{setFastPrimerTrim}
\title{Create a fastPrimerTrim object
Users have the option of trimming adapters from their data using bbduk.sh}
\usage{
setFastPrimerTrim(inDir, outDir, adapterList)
}
\arguments{
\item{inDir}{Required input. Path to directory containing FASTQ files.}

\item{outDir}{Required input. Path to write output files.}

\item{adapterList}{Required input. List of adapter sequences to remove from files.

TODO: Add Example}
}
\description{
Create a fastPrimerTrim object
Users have the option of trimming adapters from their data using bbduk.sh
}
